Add extra HasType constraint to Lam constructor.
This is necessary when writing a translator from the 'E a' type
(which has a shallow embedding of its type system in Haskell's)
to another representation using a deep embedding (i.e. data representing
the types).

The type of something constructed with the 'Lam' constructor is 'E (a -> b)'.
Unfortunately, pattern matching on this constructor will only tell us
that 'HasType a' is satisified, but not also 'HasType b'. This is why
the 'HasType b' constraint was added.